Address round 4 Copilot feedback for plan-52-openwakeword
- Extract _find_hw_input_device() to common/audio_devices.py; import it
  in wake_word.py and barge_in.py to eliminate duplication without any
  circular import
- Move _suppress_alsa_errors() above the SDL ALSA block in audio.py and
  call it before pyaudio.PyAudio() so ALSA errors are suppressed during
  device enumeration at import time
- Fix vad_recorder.py comment: was 'hold the mic for 30s' (vad_timeout),
  actual behaviour uses vad_silence_duration — clarify the distinction
- Generalise README tflite note: drop specific Python version range to
  avoid future drift
