+++ common/barge_in.py
@@ -1,34 +1,13 @@
 import logging
-import platform
-import re
 import struct
 import threading
 import time
 
+from common.audio_devices import _find_hw_input_device
 from common.openwakeword_detector import OpenWakeWordDetector
 import pyaudio
 
 logger = logging.getLogger(__name__)
-
-
-def _find_hw_input_device(pa):
-    """Return the index of the first real hardware input device on Linux.
-
-    Mirrors the same helper in wake_word.py. Kept separate to avoid a
-    circular import (wake_word imports barge_in).
-    Returns None on non-Linux platforms or when no hw: device is found.
-    """
-    if platform.system().lower() != "linux":
-        return None
-    try:
-        for i in range(pa.get_device_count()):
-            dev = pa.get_device_info_by_index(i)
-            if dev.get("maxInputChannels", 0) > 0 and re.search(r'hw:\d+,\d+', dev.get("name", "")):
-                logger.debug("Barge-in: selected hw input device index=%s name=%s", i, dev.get("name", ""))
-                return i
-    except Exception as e:
-        logger.warning("Barge-in: failed to enumerate audio devices, using default: %s", e)
-    return None
 
 # Sentinel returned by run_with_polling when barge-in interrupted the operation.
 _BARGE_IN = object()
