+++ common/wake_word.py
@@ -8,6 +8,7 @@ import time
 from enum import Enum
 
 import pyaudio
+from common.audio_devices import _find_hw_input_device
 from common.openwakeword_detector import OpenWakeWordDetector
 
 from common.beep_generator import create_confirmation_beep, create_ack_earcon
@@ -19,30 +20,6 @@ from common.utils import _is_enabled_flag
 
 logger = logging.getLogger(__name__)
 
-
-def _find_hw_input_device(pa):
-    """Return the index of the first real hardware input device (name contains 'hw:N,M').
-
-    On Linux the PyAudio default input is often a virtual 'default' device that
-    may not deliver audio from the actual USB mic. Explicitly picking the first
-    hw: input device ensures we use the real hardware.
-
-    Returns None on macOS/non-Linux (let PyAudio pick the system default).
-    Returns None (and logs a warning) if device enumeration fails.
-    """
-    import re
-    import platform
-    if platform.system().lower() != "linux":
-        return None
-    try:
-        for i in range(pa.get_device_count()):
-            dev = pa.get_device_info_by_index(i)
-            if dev.get("maxInputChannels", 0) > 0 and re.search(r'hw:\d+,\d+', dev.get("name", "")):
-                logger.debug("Wake word: selected hw input device index=%s name=%s", i, dev.get("name", ""))
-                return i
-    except Exception as e:
-        logger.warning("Wake word: failed to enumerate audio devices, using default: %s", e)
-    return None
 
 
 class State(Enum):
